Moved `PersonCreatedEvent` and related structs and tests to submodule `people::event`.
