+++ src/people/mod.rs
@@ -68,18 +68,20 @@
 //! queries. However, you force an index to be created for a single
 //! property by using [`Context::index_property()`].
 
+mod event;
 mod index;
 mod property;
 mod query;
 
 use crate::{
-    context::{Context, IxaEvent},
+    context::Context,
     define_data_plugin,
     error::IxaError,
     random::{ContextRandomExt, RngId},
     Tabulator,
 };
 
+pub use event::{PersonCreatedEvent, PersonPropertyChangeEvent};
 use index::{Index, IndexValue};
 pub use property::{
     define_derived_property, define_person_property, define_person_property_with_default,
@@ -87,7 +89,6 @@ pub use property::{
 };
 use query::Query;
 
-use ixa_derive::IxaEvent;
 use rand::Rng;
 use seq_macro::seq;
 use serde::{Deserialize, Serialize};
@@ -428,35 +429,6 @@ impl Iterator for PeopleIterator {
         self.person_id += 1;
 
         ret
-    }
-}
-
-/// Emitted when a new person is created
-/// These should not be emitted outside this module
-#[derive(Clone, Copy, IxaEvent)]
-#[allow(clippy::manual_non_exhaustive)]
-pub struct PersonCreatedEvent {
-    /// The [`PersonId`] of the new person.
-    pub person_id: PersonId,
-}
-
-/// Emitted when a person property is updated
-/// These should not be emitted outside this module
-#[derive(Copy, Clone)]
-#[allow(clippy::manual_non_exhaustive)]
-pub struct PersonPropertyChangeEvent<T: PersonProperty> {
-    /// The [`PersonId`] that changed
-    pub person_id: PersonId,
-    /// The new value
-    pub current: T::Value,
-    /// The old value
-    pub previous: T::Value,
-}
-impl<T: PersonProperty + 'static> IxaEvent for PersonPropertyChangeEvent<T> {
-    fn on_subscribe(context: &mut Context) {
-        if T::is_derived() {
-            context.register_property::<T>();
-        }
     }
 }
 
@@ -1007,7 +979,7 @@ impl ContextPeopleExtInternal for Context {
 
 #[cfg(test)]
 mod test {
-    use super::{ContextPeopleExt, PersonCreatedEvent, PersonId, PersonPropertyChangeEvent};
+    use super::{ContextPeopleExt, PersonId, PersonPropertyChangeEvent};
     use crate::{
         context::Context,
         define_derived_property, define_global_property, define_person_property,
@@ -1024,8 +996,8 @@ mod test {
         Child,
         Adult,
     }
-    define_global_property!(Threshold, u8);
-    define_derived_property!(IsEligible, bool, [Age], [Threshold], |age, threshold| {
+    define_global_property!(ThresholdP, u8);
+    define_derived_property!(IsEligible, bool, [Age], [ThresholdP], |age, threshold| {
         age >= threshold
     });
 
@@ -1072,22 +1044,6 @@ mod test {
         [AdultRunner, AdultSwimmer],
         |adult_runner, adult_swimmer| { adult_runner || adult_swimmer }
     );
-
-    #[test]
-    fn observe_person_addition() {
-        let mut context = Context::new();
-
-        let flag = Rc::new(RefCell::new(false));
-        let flag_clone = flag.clone();
-        context.subscribe_to_event(move |_context, event: PersonCreatedEvent| {
-            *flag_clone.borrow_mut() = true;
-            assert_eq!(event.person_id.0, 0);
-        });
-
-        let _ = context.add_person(()).unwrap();
-        context.execute();
-        assert!(*flag.borrow());
-    }
     #[test]
     fn set_get_properties() {
         let mut context = Context::new();
@@ -1204,54 +1160,6 @@ mod test {
     }
 
     #[test]
-    fn observe_person_property_change() {
-        let mut context = Context::new();
-
-        let flag = Rc::new(RefCell::new(false));
-        let flag_clone = flag.clone();
-        context.subscribe_to_event(
-            move |_context, event: PersonPropertyChangeEvent<RiskCategory>| {
-                *flag_clone.borrow_mut() = true;
-                assert_eq!(event.person_id.0, 0, "Person id is correct");
-                assert_eq!(
-                    event.previous,
-                    RiskCategoryValue::Low,
-                    "Previous value is correct"
-                );
-                assert_eq!(
-                    event.current,
-                    RiskCategoryValue::High,
-                    "Current value is correct"
-                );
-            },
-        );
-        let person_id = context
-            .add_person((RiskCategory, RiskCategoryValue::Low))
-            .unwrap();
-        context.set_person_property(person_id, RiskCategory, RiskCategoryValue::High);
-        context.execute();
-        assert!(*flag.borrow());
-    }
-
-    #[test]
-    fn observe_person_property_change_with_set() {
-        let mut context = Context::new();
-
-        let flag = Rc::new(RefCell::new(false));
-        let flag_clone = flag.clone();
-        context.subscribe_to_event(
-            move |_context, _event: PersonPropertyChangeEvent<RunningShoes>| {
-                *flag_clone.borrow_mut() = true;
-            },
-        );
-        let person_id = context.add_person(()).unwrap();
-        // Initializer called as a side effect of set, so event fires.
-        context.set_person_property(person_id, RunningShoes, 42);
-        context.execute();
-        assert!(*flag.borrow());
-    }
-
-    #[test]
     fn initialize_without_initializer_succeeds() {
         let mut context = Context::new();
         context
@@ -1298,26 +1206,6 @@ mod test {
             context.get_person_property(person, AgeGroup),
             AgeGroupValue::Adult
         );
-    }
-    #[test]
-    fn get_person_property_change_event() {
-        let mut context = Context::new();
-        let person = context.add_person((Age, 17)).unwrap();
-
-        let flag = Rc::new(RefCell::new(false));
-
-        let flag_clone = flag.clone();
-        context.subscribe_to_event(
-            move |_context, event: PersonPropertyChangeEvent<AgeGroup>| {
-                assert_eq!(event.person_id.0, 0);
-                assert_eq!(event.previous, AgeGroupValue::Child);
-                assert_eq!(event.current, AgeGroupValue::Adult);
-                *flag_clone.borrow_mut() = true;
-            },
-        );
-        context.set_person_property(person, Age, 18);
-        context.execute();
-        assert!(*flag.borrow());
     }
 
     #[test]
@@ -1414,7 +1302,7 @@ mod test {
     #[test]
     fn get_derived_property_with_globals() {
         let mut context = Context::new();
-        context.set_global_property_value(Threshold, 18).unwrap();
+        context.set_global_property_value(ThresholdP, 18).unwrap();
         let child = context.add_person((Age, 17)).unwrap();
         let adult = context.add_person((Age, 19)).unwrap();
         assert!(!context.get_person_property(child, IsEligible));
